Downgrade longs to ints in C# draft

+++ etc/problem201_3.cs
@@ -5,16 +5,16 @@ using System;
 using System.Collections.Generic;
 
 public class problem201_4 {
-  private static IDictionary<(long, long, long), bool> cache = new Dictionary<(long, long, long), bool>();
-  private static IDictionary<(long, long, long), bool> uniqueCache = new Dictionary<(long, long, long), bool>();
+  private static IDictionary<(int, int, int), bool> cache = new Dictionary<(int, int, int), bool>();
+  private static IDictionary<(int, int, int), bool> uniqueCache = new Dictionary<(int, int, int), bool>();
 
   public static void Main(String[] args) {
     //Console.WriteLine(canUniquelySumTo(5, 2, 2)); // True
     //Console.WriteLine(canUniquelySumTo(5, 2, 3)); // False
     //Console.WriteLine(canUniquelySumTo(89, 8, 3)); // False (non-unique)
-    long totalSum = 0;
+    int totalSum = 0;
     // 42,925 = sum of first 50 squares, 295,425 = sum of next 50 squares
-    for (long i = 42_925; i <= 295_425; i++) {
+    for (int i = 42_925; i <= 295_425; i++) {
       //if (i % 1000 == 0) {
       //  Console.WriteLine(i);
       //}
@@ -28,15 +28,15 @@ public class problem201_4 {
     Console.WriteLine(totalSum);
   }
 
-  private static long sumSquares(long n) {
+  private static int sumSquares(int n) {
     return n * (n + 1) * (2 * n + 1) / 6;
   }
 
-  private static long sumSquares(long i, long j) {
+  private static int sumSquares(int i, int j) {
     return sumSquares(j) - sumSquares(i - 1);
   }
 
-  private static bool canSumTo(long target, long maxTerm, long termCount) {
+  private static bool canSumTo(int target, int maxTerm, int termCount) {
     if (termCount == 0) {
       return (target == 0);
     } else if (termCount < 0) {
@@ -64,7 +64,7 @@ public class problem201_4 {
     }
   }
 
-  private static bool canUniquelySumTo(long target, long maxTerm, long termCount) {
+  private static bool canUniquelySumTo(int target, int maxTerm, int termCount) {
     if (uniqueCache.TryGetValue((target, maxTerm, termCount), out bool result)) {
       return result;
     }
@@ -73,14 +73,14 @@ public class problem201_4 {
     return result;
   }
 
-  private static bool canUniquelySumToImpl(long target, long maxTerm, long termCount) {
+  private static bool canUniquelySumToImpl(int target, int maxTerm, int termCount) {
     if (termCount == 0) {
       return (target == 0);
     } else if (target < 0) {
       return false;
     } else {
       bool foundOne = false;
-      for (long i = maxTerm; i >= 1; i--) {
+      for (int i = maxTerm; i >= 1; i--) {
         if (canSumTo(target - i * i, i - 1, termCount - 1)) {
           if (!canUniquelySumTo(target - i * i, i - 1, termCount - 1)) {
             // Not unique
